+++ src/wrapper/java_vm/init_args.rs
@@ -21,7 +21,7 @@ error_chain! {
 pub struct InitArgsBuilder {
     opts: Vec<String>,
     ignore_unrecognized: bool,
-    version: i32,
+    version: JNIVersion,
 }
 
 impl Default for InitArgsBuilder {
@@ -29,7 +29,7 @@ impl Default for InitArgsBuilder {
         InitArgsBuilder {
             opts: vec![],
             ignore_unrecognized: false,
-            version: JNIVersion::V1.into(),
+            version: JNIVersion::V8,
         }
     }
 }
@@ -58,10 +58,10 @@ impl InitArgsBuilder {
 
     /// Set JNI version for the init args
     ///
-    /// Default: V1
+    /// Default: V8
     pub fn version(self, version: JNIVersion) -> Self {
         let mut s = self;
-        s.version = version.into();
+        s.version = version;
         s
     }
 
@@ -96,7 +96,7 @@ impl InitArgsBuilder {
 
         Ok(InitArgs {
             inner: JavaVMInitArgs {
-                version: self.version,
+                version: self.version.into(),
                 ignoreUnrecognized: self.ignore_unrecognized as _,
                 options: opts.as_ptr() as _,
                 nOptions: opts.len() as _,